+++ src/pages/[post]/og.jpg.ts
@@ -1,12 +1,12 @@
 import type { APIContext } from "astro";
-import { getCover } from "~/image.ts";
-import { getOpenGraphImage } from "~/node.ts";
+import { getCover, getOpenGraphImage } from "~/image.ts";
 import { getEntry } from "~/site.astro";
 
 export const prerender = false;
 
 export async function GET({ params }: APIContext) {
-  const post = await getEntry("posts", params.post ?? "");
+  const post = await getEntry("posts", params.post ?? "") ??
+    await getEntry("pages", params.post ?? "");
   if (!post) return new Response("Not found", { status: 404 });
   const cover = await getCover(post.data.cover);
   return await getOpenGraphImage({
