+++ src/pages/photography/[photo]/og.jpg.ts
@@ -1,6 +1,6 @@
 import { APIContext } from "astro";
 import { formatDate } from "~/date.ts";
-import { getOpenGraphImage } from "~/node.ts";
+import { getOpenGraphImage } from "~/image.ts";
 import { getEntry } from "~/site.astro";
 
 export const prerender = false;
